+++ translations/en_GB.ts
@@ -1157,6 +1157,10 @@ Configure cache. Slides are rendered to compressed cache. These settings defined
         <source>select pdfpc/JSON file for overlays and notes</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>slide background color</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ShapeSelectionButton</name>